requirements.txt
Django==3.0.7
djangorestframework==3.11.0
psycopg2>=2.8.5, <2.9.0
Pillow>=7.1.2,<7.2.0
flake8>=3.8.3,<3.9.0
uWSGI==2.0.19.1